requirements.txt
future==0.18.2
iso8601==0.1.16
pkg-resources==0.0.0
pyserial==3.5
PyYAML==5.4.1
